// src/api/generated/model/countryUpdate.ts
/**
 * Generated by orval v6.27.1 🍺
 * Do not edit manually.
 * API Geo
 * OpenAPI spec version: 0.1.0
 */
import type { CountryUpdateDeletedAt } from "./countryUpdateDeletedAt";
import type { CountryUpdateIso3116Alpha2 } from "./countryUpdateIso3116Alpha2";
import type { CountryUpdateIso3166Alpha3 } from "./countryUpdateIso3166Alpha3";
import type { CountryUpdateLatitude } from "./countryUpdateLatitude";
import type { CountryUpdateLongitude } from "./countryUpdateLongitude";
import type { CountryUpdateName } from "./countryUpdateName";
import type { CountryUpdateNeedAutomaticUpdate } from "./countryUpdateNeedAutomaticUpdate";
import type { CountryUpdateOsmId } from "./countryUpdateOsmId";
import type { CountryUpdateOsmType } from "./countryUpdateOsmType";
import type { CountryUpdatePhoneCode } from "./countryUpdatePhoneCode";
import type { CountryUpdatePhoneMask } from "./countryUpdatePhoneMask";

export interface CountryUpdate {
  deleted_at?: CountryUpdateDeletedAt;
  iso3116_alpha2?: CountryUpdateIso3116Alpha2;
  iso3166_alpha3?: CountryUpdateIso3166Alpha3;
  latitude?: CountryUpdateLatitude;
  longitude?: CountryUpdateLongitude;
  name?: CountryUpdateName;
  need_automatic_update?: CountryUpdateNeedAutomaticUpdate;
  osm_id?: CountryUpdateOsmId;
  osm_type?: CountryUpdateOsmType;
  phone_code?: CountryUpdatePhoneCode;
  phone_mask?: CountryUpdatePhoneMask;
}